+++ NEWS.md
@@ -9,6 +9,7 @@
 - Fix bug in `PolyVtx` (#194)
 - Fix bug in feature-level subsetting (#200)
 - Update `UpdateSeuratObject` to run without `Seurat` installed (#199)
+- Add warning in `Layers.Assay()` when the search returns no results (@maxim-h, #189)
 
 # SeuratObject 5.0.1